Move adapters to "adapters" package
