+++ app/src/main/java/com/example/android/spotifystreamer/ParcelableTrack.java
@@ -4,11 +4,11 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 public class ParcelableTrack implements Parcelable {
-    String id;
-    String title;
-    String album;
-    String poster;
-    String previewUrl;
+    public String id;
+    public String title;
+    public String album;
+    public String poster;
+    public String previewUrl;
 
     @Override
     public int describeContents() {
